- Moved tests. - Updated code to work with new packages.
